+++ jvm-packages/xgboost4j-spark/src/main/scala/ml/dmlc/xgboost4j/scala/spark/DataUtils.scala
@@ -16,6 +16,8 @@
 
 package ml.dmlc.xgboost4j.scala.spark
 
+import scala.collection.mutable
+
 import ml.dmlc.xgboost4j.{LabeledPoint => XGBLabeledPoint}
 
 import org.apache.spark.HashPartitioner
@@ -99,43 +101,129 @@ object DataUtils extends Serializable {
     }
   }
 
+  /** Packed parameters used by [[convertDataFrameToXGBLabeledPointRDDs]] */
+  private[spark] case class PackedParams(labelCol: Column,
+    featuresCol: Column,
+    weight: Column,
+    baseMargin: Column,
+    group: Option[Column],
+    numWorkers: Int,
+    deterministicPartition: Boolean)
+
+  /**
+   * convertDataFrameToXGBLabeledPointRDDs converts DataFrames to an array of RDD[XGBLabeledPoint]
+   *
+   * First, it serves converting each instance of input into XGBLabeledPoint
+   * Second, it repartition the RDD to the number workers.
+   *
+   */
   private[spark] def convertDataFrameToXGBLabeledPointRDDs(
-      labelCol: Column,
-      featuresCol: Column,
-      weight: Column,
-      baseMargin: Column,
-      group: Option[Column],
-      numWorkers: Int,
-      deterministicPartition: Boolean,
-      dataFrames: DataFrame*): Array[RDD[XGBLabeledPoint]] = {
-    val selectedColumns = group.map(groupCol => Seq(labelCol.cast(FloatType),
-      featuresCol,
-      weight.cast(FloatType),
-      groupCol.cast(IntegerType),
-      baseMargin.cast(FloatType))).getOrElse(Seq(labelCol.cast(FloatType),
-      featuresCol,
-      weight.cast(FloatType),
-      baseMargin.cast(FloatType)))
-    val arrayOfRDDs = dataFrames.toArray.map {
-      df => df.select(selectedColumns: _*).rdd.map {
-        case row @ Row(label: Float, features: Vector, weight: Float, group: Int,
-          baseMargin: Float) =>
-          val (size, indices, values) = features match {
-            case v: SparseVector => (v.size, v.indices, v.values.map(_.toFloat))
-            case v: DenseVector => (v.size, null, v.values.map(_.toFloat))
+    packedParams: PackedParams,
+    dataFrames: DataFrame*): Array[RDD[XGBLabeledPoint]] = {
+
+    packedParams match {
+      case j @ PackedParams(labelCol, featuresCol, weight, baseMargin, group, numWorkers,
+      deterministicPartition) =>
+        val selectedColumns = group.map(groupCol => Seq(labelCol.cast(FloatType),
+          featuresCol,
+          weight.cast(FloatType),
+          groupCol.cast(IntegerType),
+          baseMargin.cast(FloatType))).getOrElse(Seq(labelCol.cast(FloatType),
+          featuresCol,
+          weight.cast(FloatType),
+          baseMargin.cast(FloatType)))
+        val arrayOfRDDs = dataFrames.toArray.map {
+          df => df.select(selectedColumns: _*).rdd.map {
+            case row @ Row(label: Float, features: Vector, weight: Float, group: Int,
+            baseMargin: Float) =>
+              val (size, indices, values) = features match {
+                case v: SparseVector => (v.size, v.indices, v.values.map(_.toFloat))
+                case v: DenseVector => (v.size, null, v.values.map(_.toFloat))
+              }
+              val xgbLp = XGBLabeledPoint(label, size, indices, values, weight, group, baseMargin)
+              attachPartitionKey(row, deterministicPartition, numWorkers, xgbLp)
+            case row @ Row(label: Float, features: Vector, weight: Float, baseMargin: Float) =>
+              val (size, indices, values) = features match {
+                case v: SparseVector => (v.size, v.indices, v.values.map(_.toFloat))
+                case v: DenseVector => (v.size, null, v.values.map(_.toFloat))
+              }
+              val xgbLp = XGBLabeledPoint(label, size, indices, values, weight,
+                baseMargin = baseMargin)
+              attachPartitionKey(row, deterministicPartition, numWorkers, xgbLp)
           }
-          val xgbLp = XGBLabeledPoint(label, size, indices, values, weight, group, baseMargin)
-          attachPartitionKey(row, deterministicPartition, numWorkers, xgbLp)
-        case row @ Row(label: Float, features: Vector, weight: Float, baseMargin: Float) =>
-          val (size, indices, values) = features match {
-            case v: SparseVector => (v.size, v.indices, v.values.map(_.toFloat))
-            case v: DenseVector => (v.size, null, v.values.map(_.toFloat))
-          }
-          val xgbLp = XGBLabeledPoint(label, size, indices, values, weight, baseMargin = baseMargin)
-          attachPartitionKey(row, deterministicPartition, numWorkers, xgbLp)
-      }
+        }
+        repartitionRDDs(deterministicPartition, numWorkers, arrayOfRDDs)
+
+      case _ => throw new IllegalArgumentException("Wrong PackedParams") // never reach here
     }
-    repartitionRDDs(deterministicPartition, numWorkers, arrayOfRDDs)
+
   }
+
+  private[spark] def processMissingValues(
+      xgbLabelPoints: Iterator[XGBLabeledPoint],
+      missing: Float,
+      allowNonZeroMissing: Boolean): Iterator[XGBLabeledPoint] = {
+    if (!missing.isNaN) {
+      removeMissingValues(verifyMissingSetting(xgbLabelPoints, missing, allowNonZeroMissing),
+        missing, (v: Float) => v != missing)
+    } else {
+      removeMissingValues(verifyMissingSetting(xgbLabelPoints, missing, allowNonZeroMissing),
+        missing, (v: Float) => !v.isNaN)
+    }
+  }
+
+  private[spark] def processMissingValuesWithGroup(
+      xgbLabelPointGroups: Iterator[Array[XGBLabeledPoint]],
+      missing: Float,
+      allowNonZeroMissing: Boolean): Iterator[Array[XGBLabeledPoint]] = {
+    if (!missing.isNaN) {
+      xgbLabelPointGroups.map {
+        labeledPoints => processMissingValues(
+          labeledPoints.iterator,
+          missing,
+          allowNonZeroMissing
+        ).toArray
+      }
+    } else {
+      xgbLabelPointGroups
+    }
+  }
+
+  private def removeMissingValues(
+    xgbLabelPoints: Iterator[XGBLabeledPoint],
+    missing: Float,
+    keepCondition: Float => Boolean): Iterator[XGBLabeledPoint] = {
+    xgbLabelPoints.map { labeledPoint =>
+      val indicesBuilder = new mutable.ArrayBuilder.ofInt()
+      val valuesBuilder = new mutable.ArrayBuilder.ofFloat()
+      for ((value, i) <- labeledPoint.values.zipWithIndex if keepCondition(value)) {
+        indicesBuilder += (if (labeledPoint.indices == null) i else labeledPoint.indices(i))
+        valuesBuilder += value
+      }
+      labeledPoint.copy(indices = indicesBuilder.result(), values = valuesBuilder.result())
+    }
+  }
+
+  private def verifyMissingSetting(
+    xgbLabelPoints: Iterator[XGBLabeledPoint],
+    missing: Float,
+    allowNonZeroMissing: Boolean): Iterator[XGBLabeledPoint] = {
+    if (missing != 0.0f && !allowNonZeroMissing) {
+      xgbLabelPoints.map(labeledPoint => {
+        if (labeledPoint.indices != null) {
+          throw new RuntimeException(s"you can only specify missing value as 0.0 (the currently" +
+            s" set value $missing) when you have SparseVector or Empty vector as your feature" +
+            s" format. If you didn't use Spark's VectorAssembler class to build your feature " +
+            s"vector but instead did so in a way that preserves zeros in your feature vector " +
+            s"you can avoid this check by using the 'allow_non_zero_for_missing parameter'" +
+            s" (only use if you know what you are doing)")
+        }
+        labeledPoint
+      })
+    } else {
+      xgbLabelPoints
+    }
+  }
+
 
 }
